@@ -9,7 +9,7 @@
     "devDependencies": {
         "@sveltejs/vite-plugin-svelte": "^3.0.2",
         "@tsconfig/svelte": "^5.0.2",
-        "svelte": "^4.2.12",
+        "svelte": "^5.0.0-next.95",
         "svelte-check": "^3.6.6",
         "tslib": "^2.6.2",
         "typescript": "^5.2.2",
