@@ -1,11 +1,11 @@
 <script lang="ts">
-    let count = 0;
+    let count = $state(0);
 
     const increment = () => {
         count += 1;
     };
 </script>
 
-<button on:click={increment}>
+<button onclick={increment}>
     Count: {count}
 </button>
